+++ MMM-RBB-Weather.js
@@ -30,7 +30,12 @@ Module.register('MMM-RBB-Weather', {
         dayFormat: 'ddd',
         splitCurrentTextGreater: 30,
         tableClass: 'small',
-        whiteIcons: true
+        whiteIcons: true,
+
+        // Trigger other modules
+        triggers: [
+            // eg. { day: 1, field: 'prr', value: 50, module: 'clock', hide: true }
+        ]
     },
 
     // Requires Nunjucks support added in Magic Mirror 2.2.0
@@ -90,6 +95,9 @@ Module.register('MMM-RBB-Weather', {
 
             // Update module
             this.updateDom(this.config.animationSpeed * 1000);
+
+            // Trigger other modules
+            this.triggerModules();
         }
     },
 
@@ -285,5 +293,53 @@ Module.register('MMM-RBB-Weather', {
         setTimeout(() => {
             this.loadData();
         }, this.config.updateInterval * 1000);
+    },
+
+    /**
+     * triggerModules - Trigger other modules based on weather data values.
+     * Hides the defined module(s) if the data field is lower or equal than the
+     * defined trigger value, otherwise shows the module. This behaviour is
+     * switched if the hide-Property is used in triggers.
+     */
+    triggerModules: function() {
+
+        // Do nothing if no data is available
+        if (this.weatherData === null) return;
+
+        // Iterate over defined triggers
+        for (const trigger of this.config.triggers) {
+
+            // Check if weather data is available for needed day
+            const data = this.weatherData[trigger.day];
+            if (!data) continue;
+
+            // Get module(s) by class
+            const modules = MM.getModules().withClass(trigger.module);
+            for (const modul of modules) {
+                Log.info(`Trigger module '${modul.name}' ...`);
+
+                // Show/hide options
+                const lockOptions = { lockString: this.name };
+                const speed = this.config.animationSpeed * 1000;
+
+                // Get weather data to check
+                let dataValue = data[trigger.field];
+                if (trigger.field === 'maxtemp') {
+                    dataValue = data.temp.split(';')[0];
+                }
+                if (trigger.field === 'mintemp') {
+                    dataValue = data.temp.split(';')[1];
+                }
+
+                // Check trigger condition: Hide module if data value is lower
+                // or equal than trigger value or if the trigger uses hide mode,
+                // but not both (xor)
+                if ((dataValue <= trigger.value) !== (trigger.hide === true)) {
+                    modul.hide(speed, lockOptions);
+                } else {
+                    modul.show(speed, lockOptions);
+                }
+            }
+        }
     }
 });