+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Optionally show update time
+- Define triggers to show/hide other modules based on weather data
 
 ### Updated
 - eslint packages (dev-dependencies) and resulting new linting errors/warnings